% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_genes.r
\name{compute_euler_plot}
\alias{compute_euler_plot}
\title{Compute and draw euleur plot based on significant DEG table}
\usage{
compute_euler_plot(
  deg_table_in,
  min_to_displayed_threshold = 0.01,
  to_keep = NULL,
  ext_groups = NULL,
  mode_ext_groups = "union",
  inter_batch = NULL,
  mode_inter_batch = NULL,
  shape = "ellipse",
  sort_group_for_region_naming = TRUE,
  path_dir_to_save_plot = NULL,
  dict_uniprot_to_symbol = NULL
)
}
\arguments{
\item{deg_table_in}{Recquired. dt with significant results: results of \code{data.table(deseq_pred$cross_args_and_generate_lists())}. (Note: differente, pvalue, log_FC and deregulaton type has been managed separatly !). Note: the “type” column should not contain “NA”, otherwise false warnings that some 'to_keep' or 'ext_group' groups/batches will be incorrectly issued. 'group' and 'abtch' must not contains '&'.}

\item{min_to_displayed_threshold}{(default: '0.01') Minimal values displayed. If <1: in proportion of universe (sum of all sets lengths (with redondancy) - including potentials ext_groups).}

\item{to_keep}{(default: NULL) dt or df with 2 column: 'batch' and 'group'. If 'NULL', no filter, keep all data. Impact inter AND intra batch comparison but no impact 'ext_group'.}

\item{ext_groups}{(default: NULL) list of uniprot ids used as ext_group (list or vector). If table (dt or df with 2 column: 'batch' and 'group') determine the gene list from 'deg_table_in' (before filtering using 'to_keep' and considering appropriate 'type', 'lfc' and 'pval'). If 'NULL', no external set. Used for inter AND intra batche comparison.}

\item{mode_ext_groups}{(default: 'union') Used only if 'ext_groups' is table. Must be 'individual', 'union' or 'intersect'.}

\item{inter_batch}{(default: NULL) list of named list with batch to compare together. If NULL, and 'mode_inter_batch' not NULL, include all batche. Performed independantly for each 'type', 'pval' and 'lfc'. Names must not contains '&'.}

\item{mode_inter_batch}{(default: NULL) list of named list (same name as 'inter_batch') specifying mode of selection of deg to each batch for each inter-batch comparison ('union' or 'intersect'). If 'NULL', no Inter-batch comparison.}

\item{shape}{(default: 'ellipse') euler diagram shape.}

\item{sort_group_for_region_naming}{(default=TRUE) Alphanumerically sorts the names of the groups composing each region into the name of the region, or concervates the order of the various groups in the batch in question as in the initial input}

\item{path_dir_to_save_plot}{(default: NULL) directory path where save individual euler diagram on separate pdf. If 'NULL', no saving.}

\item{dict_uniprot_to_symbol}{(default: NULL) used to reconstitude column 'symbol' of table in 'data' of each result if 'symbol' column is already present in 'deg_table_in' input (if 'symbol column not in input table, this parameters will be ignored). If NULL, reconstructed from input table.}
}
\value{
This function return as first element a list of eullerr plots object ('$euler_plot') and as second element a data.table with specific content of each region on format of 'deg_table_in' ('$dt_region_content').
}
\description{
Computed at uniprot level ! 'NA' value has been ignored. (quid of multiple uniprot_id for a single gene_symbol ?!?!?!)
}
\details{
Comparison is performed using uniprot ids.

TODO:
\itemize{
\item in option management of up/down. Although this is already possible in the current state of development by transforming deg_table upstream, it would be a good idea to add the following 2 options)
\itemize{
\item summarize analyse: merge up and down on single analyse (after prefixing genes by their status "up_"/"down_")
\item very interesting analyse: venn with up and down in same plot on disctinct sets to highlight swtched genes !
}
\item ext_groups:
\itemize{
\item add possibility to put mutliple specific gene list as vector (with mode: union, individual, intersect)
\item add option to use ext group at inter AND/OR intra batche comparison (actually one sigle group used for both)
}
\item plot:
\itemize{
\item put all venn diagram on same pdf using output/append on new page option !
\item manage tittle parameters
}
\item reduce time running: don't perform independantly euleurr plot ! use already computed intersection to put stats to eulleur diagram
}
}
\examples{
Example usage

compute_euler_plot(deg_table = data.table(deseq_pred$cross_args_and_generate_lists(
                                            cross_type = c("deregulated"),
                                            cross_lfc_abs_lim = c(LFC_threshold),
                                            cross_min_signif = c(pvaladj_threshold)
                                         )
                  )
                  
}
