% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_genes.r
\name{draw_single_euler_plot}
\alias{draw_single_euler_plot}
\title{Draw simple and single euler plot about a named list of sets}
\usage{
draw_single_euler_plot(
  sets,
  min_to_displayed_threshold = 0.01,
  shape = "ellipse",
  title = NULL,
  save_path_file = NULL
)
}
\arguments{
\item{min_to_displayed_threshold}{Minimal values displayed. If <1: in proportion of universe (sum of all sets lengths (with redondancy)).}
}
\description{
Draw simple and single euler plot about a named list of sets
}
