% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_genes.r
\name{extract_venn_zones_content}
\alias{extract_venn_zones_content}
\title{Extract specific content of each zone of a Venn/Euler diagram}
\usage{
extract_venn_zones_content(sets_list, path_xlsx_file_to_save_table = NULL)
}
\arguments{
\item{sets_list}{A named list of sets or a list of named lists of sets. Each list will be processed independently.}

\item{path_xlsx_file_to_save_table}{File path to save results in an Excel file (default: NULL). If not NULL, creates one sheet per set.}
}
\value{
A named list of results for each input set or list of sets. Each result contains specific content of each zone.
}
\description{
Extract specific content of each zone of a Venn/Euler diagram
}
\examples{
set123 <- list("111" = c("12345", "23456", "34567"),
               "222" = c("23456", "45678"),
               "333" = c("12345", "45678", "56789"))
setABC <- list("AAA" = c("ABCDE", "BCDEF", "DEFGH"),
               "BBB" = c("BCDEF", "DEFGH"),
               "CCC" = c("DEFGH", "EFGHI", "FGHIJ"))
extract_venn_zones_content(list("123" = set123,
                                "ABC" = setABC),
                           "test.xlsx")

}
