% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fctBio.r
\name{fctBio}
\alias{fctBio}
\title{fctBio: Objects and methods for functional biology analysis}
\description{
This package provides ...
}
\details{
Those steps are specified in the schematic below with the names of the
different class.
}
