% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ann_space.r
\name{load_ann_space}
\alias{load_ann_space}
\title{Load annotation space}
\usage{
load_ann_space(
  ann_sources,
  ann_types,
  compute_parent_genes = TRUE,
  use_parents_union = TRUE
)
}
\arguments{
\item{ann_sources}{named character vector with database code and database
path (see \link{enrich}) ann_source arg}

\item{ann_types}{named character vector with database code as keys and
database types to keep (see \link{enrich}) type_keep arg. Default is to
keep all.}

\item{compute_parent_genes}{whether to compute or not parent genes}
}
\value{
a data.table
}
\description{
Load annotation space
}
