% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_genes.r
\name{query_genes}
\alias{query_genes}
\title{Query expression tags}
\usage{
query_genes(
  nested_df,
  filter_name,
  value,
  genes_per_clusters = NULL,
  batch_labels = NULL,
  group_labels = NULL,
  data_col = "data",
  data_nested_id = "uniprot",
  enrich_gene_col = "genes"
)
}
\arguments{
\item{nested_df}{nested tibble with
\itemize{
\item gene sets given in a tibble in \code{data_col} variable and in \code{data_nested_id}
nested variable
\item eventually enrichment data in the enrich column
from the filter_and_get_nested_results method of NestedEnrich object
}}

\item{filter_name}{filter to apply
\itemize{
\item batch : select genes if present in batch(s)
\item batch_label : select genes if present in batch label(s)
\item group : select genes if present in group(s)
\item group_label : select genes if present in group label(s)
\item term_id : select genes if part of a term based on the \code{enrich_gene_col}.
\item term_name : select genes if part of a term based on the \code{enrich_gene_col}.
\item cluster : select genes if part of a cluster (significant term).
\item intra_occurences : select genes based on their minimal occurences in terms
intra cluster
\item extra_occurences : select genes based on their minimal occurences in terms
\item cluster_occurences : select genes based on their minimal occurences in
cluster
}}

\item{value}{value or vector of value to be filtered on based on filter_name}

\item{genes_per_clusters}{from the count_gene_per_cluster method of NestedEnrich object with columns:
\itemize{
\item "ID"
\item "Cluster"
\item "# in enriched term intra"
\item "# in batch"
\item "# in group"
\item "# in enriched term"
\item "# in cluster"
}}

\item{batch_labels}{batch labels}

\item{group_labels}{group labels}

\item{data_col}{see \code{nested_df}}

\item{data_nested_id}{see \code{nested_df}}

\item{enrich_gene_col}{coloum to select genes (default is genes)
\itemize{
\item GI for gene of intermterest
\item genes for all genes in terms
}}

\item{data_univ_col}{same as \code{data_col} but for the "universe"}

\item{enrich_col}{column for the nested enrichment data. Default is "enrich".
Is set to null if not in nested_df}
}
\description{
Function to build a query to find a list of genes.
}
