% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_genes.r
\name{query_genes_choices}
\alias{query_genes_choices}
\title{Get possible choices for a query}
\usage{
query_genes_choices(
  nested_df,
  filter_name,
  genes_per_clusters = NULL,
  batch_labels = NULL,
  group_labels = NULL
)
}
\arguments{
\item{nested_df}{nested tibble with
\itemize{
\item gene sets given in a tibble in \code{data_col} variable and in \code{data_nested_id}
nested variable
\item eventually enrichment data in the enrich column
from the filter_and_get_nested_results method of NestedEnrich object
}}

\item{filter_name}{filter to apply
\itemize{
\item batch : select genes if present in batch(s)
\item batch_label : select genes if present in batch label(s)
\item group : select genes if present in group(s)
\item group_label : select genes if present in group label(s)
\item term_id : select genes if part of a term based on the \code{enrich_gene_col}.
\item term_name : select genes if part of a term based on the \code{enrich_gene_col}.
\item cluster : select genes if part of a cluster (significant term).
\item intra_occurences : select genes based on their minimal occurences in terms
intra cluster
\item extra_occurences : select genes based on their minimal occurences in terms
\item cluster_occurences : select genes based on their minimal occurences in
cluster
}}

\item{genes_per_clusters}{from the count_gene_per_cluster method of NestedEnrich object with columns:
\itemize{
\item "ID"
\item "Cluster"
\item "# in enriched term intra"
\item "# in batch"
\item "# in group"
\item "# in enriched term"
\item "# in cluster"
}}

\item{batch_labels}{batch labels}

\item{group_labels}{group labels}
}
\description{
get choices to query genes (see query_genes function)
}
