% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_NestedEnrich.r
\name{enrich}
\alias{enrich}
\title{Db enrichment}
\usage{
enrich(
  gene_input,
  ann,
  gene_universe = NULL,
  lim_pmin = 0.05,
  classic = FALSE,
  log_level = "WARN",
  hard_pmin_filter = TRUE
)
}
\arguments{
\item{gene_input}{input gene list as INTEGER vector}

\item{ann}{tibble / annotation table.
term: id of the term using integer;
genes: nested list with interger representing genes.
type: The type this term fits in. To allow the subsetting of big functional
database like gene ontologye where there is 3 type: biological_process,
molecular_function and cellular_component
parents: comma separated list of parent terms (with id).}

\item{gene_universe}{universe named gene list as INTEGER vector.}

\item{lim_pmin}{minimum value for pmin (lowest possible p-value).
This limit is used for filtering before pval correction}

\item{classic}{Do analysis relative to the universe (vs parent-child)}

\item{log_level}{Level of logging (see logging package). Default = WARN}

\item{gene2int}{A dictionary to transform gene to integer.}

\item{regex}{regex pattern to remove from gene name from input lists (useful
for deleting accession version info which should not be present on database
file)}

\item{type_keep}{character vector of term type to keep (keep all by default);
e.g Biological_process for gene ontology}
}
\description{
R program designed to perform an enrichment analysis with topology-based
annotation.
Over-representation analysis is performed according to the
parent-child-intersection method.
Multi-testing is done with BH and Bonferoni methods on the subset of term
with a minimal p-value below 1e-4 by default.
P-values are based on combination analysis as described in the parent-child
paper using GNU multiple precision arithmetic
libraries and their R implementation.
}
