% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysbio_genes.r
\name{extend_trrust}
\alias{extend_trrust}
\title{Extend a list of genes by adding linked transcription factors (TFs) and/or target genes.}
\usage{
extend_trrust(
  trrust_db = system.file("extdata/human_network_annotations/trrust/data.tab", package =
    "fctBio"),
  input_genes,
  add_tf = TRUE,
  add_target = TRUE,
  keep_input_without_hit = TRUE,
  ignore_single_pmid = FALSE,
  detailled_output = FALSE
)
}
\arguments{
\item{trrust_db}{Path to the TRRUST database in tsv format.}

\item{input_genes}{A character vector of input genes.}

\item{add_tf}{Logical, whether to add transcription factors regulating the input genes (keeping 'tf' already present in input genes).}

\item{add_target}{Logical, whether to add target genes regulated by the input genes (keeping 'target' already present in input genes).}

\item{keep_input_without_hit}{Logical, whether to keep all input genes (including 'tf', 'target' and unannotated gene).}

\item{ignore_single_pmid}{Logical, whether to ignore relations supported by only one PMID reference.}

\item{detailled_output}{\link{not implemented} To return in addition of vector of gene names, a dataframe with relation of each genes pairs.}
}
\value{
A character vector of extended genes.
}
\description{
Extend a list of genes by adding linked transcription factors (TFs) and/or target genes.
}
\examples{
extended_genes <- extend_trrust(
  trrust_db = "path/to/trrust_rawdata.human.tsv",
  input_genes = c("BAX", "MYC", "TP53"),
  add_tf = TRUE,
  add_target = TRUE,
  keep_input_without_hit = TRUE,
  ignore_single_pmid = FALSE
)
}
