% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ann_space.r
\name{load_ann_space}
\alias{load_ann_space}
\title{Load Annotation Space}
\usage{
load_ann_space(
  ann_sources,
  ann_types,
  compute_parent_genes = TRUE,
  use_parents_union = TRUE
)
}
\arguments{
\item{ann_sources}{Named character vector with keys as database codes and
values as paths to tabular files containing annotations.
Each file must include the following columns:
\itemize{
\item \code{term}: Short identifier of the term.
\item \code{genes}: Semicolon-separated list of gene identifiers associated with
each term.
\item \code{type}: Character string defining the category of the term, e.g.,
Biological Process, Molecular Function, or Cellular Component in the
context of Gene Ontology.
\item \code{name}: Descriptive name or long identifier of the term.
\item \code{parents}: Comma-separated list of parent terms' identifiers, reflecting
the hierarchical structure in databases using DAGs to organize terms.
}}

\item{ann_types}{Named character vector with keys as database codes and
values specifying
the types of terms to retain from the annotations. The default behavior is to
keep all types. This parameter filters terms based on their \code{type} field.}

\item{compute_parent_genes}{Logical, indicating whether to compute a
collective list of genes for each term by incorporating genes from its parent
terms. Defaults to TRUE.}

\item{use_parents_union}{Logical, indicating how to handle gene lists from
parent terms:
\itemize{
\item If TRUE, computes the union of genes across all parent terms.
\item If FALSE, computes the intersection of genes across all parent terms.
}}
}
\value{
A data.table object structured as follows:
\itemize{
\item \code{term}: Short identifier of the term.
\item \code{name}: Long name or descriptive identifier of the term.
\item \code{genes}: List of gene identifiers associated with each term.
\item \code{parents}: List of parent terms' identifiers.
\item \code{parent_genes}: Computed list of genes derived from parent terms, present
only if
\code{compute_parent_genes} is TRUE.
\item \code{type}: Category of the term, consistent with the \code{type} field in the
input files.
\item \code{ann_name}: Database code from which each term was sourced, aligning with
keys in \code{ann_sources}.
}
}
\description{
This function loads gene annotations from specified sources, such as
databases or files, and processes them into a structured format suitable for
further analysis.
Annotations typically define the notion of "terms" which can represent
biological pathways, communities of genes in coexpression networks, or other
relevant biological groupings. These terms might be organized hierarchically
using directed acyclic graphs (DAGs) in systems like Gene Ontology or
Reactome, allowing for the representation of complex biological
relationships.
}
