% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_genes.r
\name{query_genes}
\alias{query_genes}
\title{Query Expression Tags for Gene Selection#'}
\usage{
query_genes(
  nested_df,
  filter_name,
  value,
  genes_per_clusters = NULL,
  batch_labels = NULL,
  group_labels = NULL
)
}
\arguments{
\item{nested_df}{A nested tibble as returned by the \code{get_results} method of the
\code{NestedEnrich} class, with \code{verbose = TRUE} to ensure descriptive column
names which facilitate easier querying.}

\item{filter_name}{A character string specifying the attribute to filter on.
Supported filters include:
\itemize{
\item \code{batch}: Select genes present in specified batch(es).
\item \code{batch_label}: Select genes based on batch label(s).
\item \code{group}: Select genes present in specified group(s).
\item \code{group_label}: Select genes based on group label(s).
\item \code{type}: Select genes present in specified type(s).
\item \code{term_id}: Select genes associated with a specific term ID(s).
\item \code{term_name}: Select genes associated with a specific term name(s).
\item \code{cluster}: Select genes associated with a specific cluster(s).
\item \code{intra_occurences}: Select genes based on their minimal occurrences within
the same term.
\item \code{extra_occurences}: Select genes based on their minimal occurrences across
all terms.
\item \code{cluster_occurences}: Select genes based on their minimal occurrences within
specified clusters.
}}

\item{value}{The value(s) to filter by, corresponding to the \code{filter_name}.
This can be a single value or a vector of values depending on the filter context.}

\item{genes_per_clusters}{Optional but required for xxx_occurrences filters
a data frame from the \code{count_gene_per_cluster}#' method in the \code{NestedEnrich}
class, which includes gene occurrence data across clusters.}

\item{batch_labels}{Optional; a vector of batch labels if \code{batch_label} filter
is used. This helps map user-friendly labels to batch identifiers in the dataset.}

\item{group_labels}{Optional; a vector of group labels if \code{group_label} filter
is used. This helps map user-friendly labels to group identifiers in the dataset.}
}
\value{
A character vector with gene IDs
}
\description{
\code{query_genes} allows for dynamic querying of a nested tibble containing gene
sets and enrichment data. This function can filter genes based on various
attributes such as batch presence, group membership, cluster occurrence, or
specific terms within an enrichment analysis. t is typically used
to extract specific subsets of genes from results generated by the \code{NestedEnrich}
class, particularly useful for further analysis or visualization.
}
\examples{
# Assuming `nested_df` is a dataset obtained from NestedEnrich with verbose = TRUE
genes_in_batch1 <- query_genes(nested_df, filter_name = "batch", value = "batch1")
genes_in_termGO <- query_genes(
  nested_df, filter_name = "term_id", value = "GO:XXXXXX"
)
genes_by_occurrence <- query_genes(nested_df, filter_name = "cluster_occurences",
                                   value = 5, genes_per_clusters = genes_cluster_data)
}
